LPS-72858 Move method getAbsolutePath to SourceUtil
